requirements.txt
flask
flask-mysqldb
pyyaml
opencv-python
tensorflow